[WIP]bad_connector_woocommerce:Common mapper for product and template
