+++ bad_connector_woocommerce/models/product/common.py
@@ -53,59 +53,6 @@ class ProductProduct(models.Model):
                 self.woo_bind_ids.backend_id.update_stock_inventory
             )
 
-    def _get_attribute_id_format(self, attribute, record, option=None):
-        """Return the attribute and attribute value's unique id"""
-        if not option:
-            return "{}-{}".format(attribute.get("name"), record.get("id"))
-        return "{}-{}-{}".format(option, attribute.get("id"), record.get("id"))
-
-    def _get_product_attribute(self, attribute_id, record, env):
-        """Get the product attribute that contains id as zero"""
-        binder = env.binder_for("woo.product.attribute")
-        created_id = self._get_attribute_id_format(attribute_id, record)
-        product_attribute = binder.to_internal(created_id)
-        if not product_attribute and not attribute_id.get("id"):
-            product_attribute = self.env["woo.product.attribute"].create(
-                {
-                    "name": attribute_id.get("name"),
-                    "backend_id": env.backend_record.id,
-                    "external_id": created_id,
-                    "not_real": True,
-                }
-            )
-        return product_attribute
-
-    def _create_attribute_values(
-        self, options, product_attribute, attribute, record, env
-    ):
-        """Create attribute value binding that doesn't contain ids"""
-        binder = env.binder_for("woo.product.attribute.value")
-        for option in options:
-            created_id = self._get_attribute_id_format(attribute, record, option)
-            product_attribute_value = binder.to_internal(created_id)
-            if not product_attribute_value:
-                attribute_id = self._get_attribute_id_format(attribute, record)
-                binder = env.binder_for("woo.product.attribute")
-                product_attr = binder.to_internal(attribute_id, unwrap=True)
-                attribute_value = self.env["product.attribute.value"].search(
-                    [
-                        ("name", "=", option),
-                        ("attribute_id", "=", product_attr.id),
-                    ],
-                    limit=1,
-                )
-                self.env["woo.product.attribute.value"].create(
-                    {
-                        "name": option,
-                        "attribute_id": product_attribute.odoo_id.id,
-                        "woo_attribute_id": product_attribute.id,
-                        "backend_id": env.backend_record.id,
-                        "external_id": created_id,
-                        "odoo_id": attribute_value.id if attribute_value else None,
-                    }
-                )
-        return True
-
 
 class WooProductProduct(models.Model):
     """Woocommerce Product Product"""
