+++ bad_connector_woocommerce/views/product_template_view.xml
@@ -73,11 +73,21 @@
                                 <group>
                                     <field name="backend_id" widget="selection" />
                                     <field name="external_id" />
+                                    <field name="status" />
+                                    <field name="tax_status" />
+                                    <field name="woo_product_qty" />
                                 </group>
                                 <group>
-                                    <field name="woo_product_qty" />
+                                    <field name="stock_status" />
+                                    <field name="price" />
+                                    <field name="regular_price" />
                                     <field name="stock_management" />
                                 </group>
+                                <label
+                                    for="woo_product_categ_ids"
+                                    string="WooCommerce Product Categories"
+                                />
+                                <field name="woo_product_categ_ids" />
                                 <label
                                     for="woo_attribute_ids"
                                     string="WooCommerce Attributes"
